Factor out duplicate code in GenericProgramRunner subclasses
